services/summary/requirements.txt
python-dotenv==1.1.0
confluent-kafka==2.9.0
openai==1.70.0
